ingestion/requirements.txt
beautifulsoup4
pycountry
openai
tqdm
python-dotenv
requests
